DL1JBE patch to nhash.c to fix execution of 64-bit Linux systems.
git-svn-id: svn://svn.berlios.de/wsjt/trunk@1794 b3ca0830-8508-0410-8253-dac51d8a7955

+++ nhash.c
@@ -174,7 +174,7 @@ acceptable.  Do NOT use for cryptographic purposes.
 #ifdef STDCALL
 uint32_t __stdcall NHASH( const void *key, size_t *length0, uint32_t *initval0)
 #else
-uint32_t nhash_( const void *key, size_t *length0, uint32_t *initval0)
+uint32_t nhash_( const void *key, int *length0, uint32_t *initval0)
 #endif
 {
   uint32_t a,b,c;                                          /* internal state */
